Fixed some navigational issues
