+++ src/Components/NavBar/NavBar.css
@@ -1,0 +1,12 @@
+#nav {
+  position: fixed;
+  bottom: 0;
+  display: flex;
+  width: 100%;
+  height: 50px;
+}
+
+#nav button {
+  flex: 1;
+  height: auto;
+}
