+++ src/Screens/ProfileScreen/ProfileScreen.css
@@ -1,0 +1,12 @@
+#profile-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+}
+
+#profile-header .profile-header-user {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+}
